Made DynamoDbTicketRegistryFacilitatorTests nested
